Capitalization
Because if there's one thing that is wrong with our code right now, it's
the fact shit isn't capitalized

- All names (aka shown when right-clicked or moused over) have been
capitalized. It's not a "wooden chair" but a "Wooden Chair" for
instance. Mostly for style and since a fair amount of objects ARE
already capitalized this way
- Fixed a few oddities here and there

As far as functionality goes there shouldn't be any problems, but it
might be controversial