Lower Supermatter Shard resistance
- Shard now detonates far faster. It still gives ample time to pull out
if cooled about correctly, but a shard going critical is now a more
serious matter

That's pretty much it. Supermatter might be made even more dangerous as
time goes on, but for now that should already be good enough. You
shouldn't have 10 minutes to fix the shard, especially seeing how
"fixing it" is as easy as dragging the shard to space