Changing how Meteor works
- Instead of constantly ramming the station, meteors now come in waves
depending on time and a random factor
- Wave delay now stands at 20 to 45 seconds. Allows the crew to do some
quick patching during this time
- Meteors now strike a bit faster and harder
- Shuttle now takes a full hour to arrive. The goal is to survive until
then. It might sound easy, but as the round ramps up you might end up
overwhelmed. Not setting up defenses is allowing the RNG to pick away as
far as right into the Bridge by the time the shuttle docks