Derelict Revamp Part One - MoMMI Nest, Atmospherics, Bridge
This is part of a massive content overhaul to the Derelict Z-level. The
goal is simply to turn the ruins into ACTUAL ruins of a station that
could have been instead of some empty husk with loadse broken floor
tiles

- Slight update to the DJ Station, mostly changing external plating into
catwalk
- MoMMI Nest now has a higher degree of completion. If you can into
wiring, it's basically plug and play
- Solars changed. Much fancier, much bigger, a lot more unfinished. It
can power some stuff as is, but if you're willing to use the solar pack
and spend some time on it, it will allow you to power some of the
Derelict (right now all of it, but relevant machines weren't added in
yet)
- Derelict Atmospherics now has a pre-made but broken distro line set up
and has all basic machinery needed for atmosian (Heater, Freezer,
Dispensers, Tank Storage Units, Portable Scrubbers and Pumps)
- Derelict Control Room is now basically a Bridge. All consoles that
would have been needed are in, except for the Communications Console to
avoid massive griffing
- The room next to that has been cut up into a Meeting Room
- The part between the MoMMI Nest and the Control Room has become some
sort of mini-Engineering, complete with a private E.V.A storage
- There's now more hints of general wreckage around those parts

I might add in more items and objects, but for now this is mostly the
barebones remake of those rooms. Above all, I'm trying to keep a basic
sense of the station being a Derelict, although I still have a lot of
progress to do