Project Asteroid Field - A new Z-level to explore and metagame in
Today, the station's Miners and Scientists go to that good old hulking
asteroid, full of minerals and artifacts just waiting to be salvaged and
used to build golden rooms and machinery of doom. However, few know that
Nanotrasen tried to extend its mining operations in the sector

The Asteroid Field is a new Z-level (Z-7). It acts as a new full-sized
map to get lost in and can potentially be used for other purposes. In
lore, Nanotrasen tried to set up a Mining Substation but encountered...
adversity

Contents :

- One big asteroid around the middle. Not nearly as big as the Mining
Asteroid one. Two other big asteroids hulk nearby
- A large amount of smaller rocks scattered around the area
- The middle asteroid was dug into to set up a Mining Substation, but it
did not turn out well
- All the building supplies are still here, but with a broken shuttle
and no teleporter, good luck doing anything with them

Right now the draft is very basic. The buildings could be fleshed out a
bit more, but for now this is what it's going to look like.