Add in the Meteor Monitoring Computer. Other changes to Meteors
Miscellaneous changes :

- Meteor Supplies code now split from the main Meteor gamemode and
called via a proc. This allows to declutter the main file significantly
and make changes to this more modular
- Change required_players to required_players_secret. Still 20 players.
This is something I mean to change on all other gamemodes sometimes
later, but if an admin forces the gamemode, let it be forced
- Fix meteor warning being mute half of the time, caused by rogue
indentation levels
- Fix failed bold tags in the meteor shuttle justification
- Inflate number of meteors sent in one wave from 100-200 to 200-500.
This goes along with OTHER changes below
- Prepare the ground for meteor events. Said events have yet to be
included
- Allow rare event at any time during Meteors, even BEFORE the main
wave, with probability 0.02/100 (so basically that one stray meteor)
- Split up meteor delay into "preinfo-delay" and "postinfo-delay". Those
times are still random, but base is 15 seconds pre-info and 45 seconds
post-info. Info arrives between both, meteors arrive at the end of the
post-info delay. Overall, this doubles wave delay (1 minute average
instead of 30 seconds average)
- Admins are warned precisely of meteor wave size and direction during
the info-dump, to allow them to check out the damage and generally
answer to deadchat banter
- In light of wave delay increase, make "small meteor waves" last longer
(2 to 5 minutes for big waves, 1 to 2 minutes for small waves and thus
not changed). Small meteor waves now only have a 10 seconds delay from
announcement to strike and big waves have a 30 seconds delay from
announcement to strike

Meteor Monitoring Computer :

- Add in the Meteor Monitoring Computer. Basically the Bhangometer of
Meteor Cataclyms
- This computer can be built via R&D with 2000 units of gold (along with
glass and sulfuric acid) and Programming 3, Engineering 2 and Bluespace
2
- A "meteor monitor building kit" is provided in place of the lone
meteor shield during Meteors supply dump. This is basically everything
you need for a computer and the board, tools not included
- This monitor will give thorough information about procedures to take
during Meteors rounds and will give thorough information about incoming
meteor waves
- Meteor waves now have a name. This is 100 % fluff, but does add a bit
of flavor
- Otherwise, the computer will give wave delay (static), number of
objects and direction of wave
- The computer does not log past meteor waves currently (mainly because
I just can't figure out how to do that cleanly)

This is a pretty thorough PR, and can clearly still be improved upon.
Don't mind the commented code, it's another thing that will not be part
of this PR