Switch Serve and Protect directives in NT Default
Note : As far as the code part goes, this is about as simple as it gets.
What is more important is if it truly is a good idea or not

- NT Default Law 2 is now Protect
- NT Default Law 3 is now Serve
- Weed out snowflake commented Law 5 from the code

My personal reasoning for this change :

- NT Default currently asks as THE validhunter lawset by excellence,
doubling as a Captain one-human law whenever needed. There's no reason
to not switch to that lawset whenever availaible, unless the Captain is
rogue (which usually only occurs if someone illegally instates himself
as the Captain)
- The Captain can literally ask for the silicons to murder everyone but
him and they have to comply. How NT would allow such a possibility is
perplexing, at best
- Since the Captain cannot be a starting antagonist (at least
currently), this GUARANTEES the AI only answers to non-antags. It is the
only lawset that does so (outside of ridiculous "[antags] are not human"
freeforms)
- The lawset generally puts ridiculously little value to human life.
You'd think a lawset where a superior can have one of his department
workers slain for his amusement would be considered a bit risque for a
station full of psychopaths

Discussion goes below