Table connection redo.
Changes tables from sets of snowflake sprites to a collection of component icons used to build the table itself.
This massively increases sprite consistency and removes the requirement of all the special-condition sprites that were missing.
Goodbye oldcode.

Only done as a Proof of Concept on reinforced tables.
The sprites are in no way final, since the shading is awful.
No other table types have been converted.