Tweaks to the map
- Add a lightswitch to the Kitchen
- Move the Bartender's lightswitch to his backroom
- Add a kitchen sink in the Bar counter area

- Modified Mech Bay area. It's now comprised of an big shutter "door"
with a small one-tile electrified window on both sides. This makes it
more of a "secure area for mechs to be stationed and maintained" area
and less of a "muh clay" area. There's currently no outside access, you
need to pass through R&D and Robotics to get in here (unless the
shutters are open)