Remote control code module
Added a system for contextual attack-selfing on items.
This on its own permits hooks into various uses, but the one majorly supported is buttons.
You could stick buttons on anything. Stick a button on a PDA to turn it on or off. Stick a button on your jetpack to toggle it on or off, to toggle its stabilising, to toggle its breathing, etc.
The datum is not tied to the remote item, which is just an example of how you would implement its functionality.

Buttons:
 - Remote datums allow the use of buttons, as well as adding them or removing them
 - Buttons have an on_press functionality
 - This allows you to make remotes for anything using, for example, a signaller button (space garage doors ho!)
 - No truly functional buttons exist - all the current ones just play sounds, for testing purposes
 - Buttons adapt to their remote - the remote passes the shape a button icon has to be, and its displacement
 - New buttons follow the base shape for remotes, and then can be attached anywhere