Adds CE's new toolbelt to the premium section of the derelict vender.
For an item so good there should be more than one in the entire universe.
I added one of these bad boys to the vender on the derelict and for bus. (Back when we had that asteroid build bus where we had to leave the station, one of these venders would have helped.)
Hopefully it won't be too OP, considering it is in the middle of space, unpowered, and requires a coin.