Remove all of the three mechfab designs that did not need research
For some reason, someone decided that out of all the designs in the Mech
Fabricator, those three designs in particular could be fabricated
without the needed base research

This is especially aggravating for the mech-mounted syringe launcher
because it is basically a research nuke and instantly fills a few
research categories (Bio, EMP, Data)

Now you'll have to research for those items like literally every single
other design on the Mech Fabricator. The Mech Sleeper is Bio 1, the
Phazon Phase Array needs Phazon. The research for the Mech-Mounted
Syringe Launcher is akin to the one needed for the Rapid Syringe
Launcher