Adds a snack called "Old Empire Bar" in reference to Age of Empires.
A few days ago the IC chat wasn't sanitized and I accidentally played `rogan.mp3`, I thougt that it was a 1 second audio file but it was a 9 minute audio file of "Rogan?" repeating over and over.
Worked on this for the last hour but I am finally at a stage where I am confident at making a pull request.

Signed-off-by: Jeroen de Neef <jeroen52@gmail.com>