Adds a crew monitoring console to medbay lobby (this is the only one that exists outside the paramedics room).
Swaps the grinder and shotgun table in the bar backroom.

Deletes a lightswitch in the research outpost breakroom that was obscured and lets face it, every single room doesn't need a light switch.  Even if the lights are switched off, the light from the hallway keeps it illuminated anyway.

Tests all my shit.

Finishes my shit.