Fix html not loading currently
There was an issue due to the fact the mapname I was using was a directory name, which browse_rsc() ignores all for the last portion of, and the js was trying to read the full name with the directory.

This makes it just use the nameShort of the map which has no directory which will be truncated.