Fix test_box, make test_box primary repo map
This both fixes the compile to test box which was broken by the removal of the celtic board.

This also makes test_box the initial map ticked by vgstation.dme, since its basically box but just better in every way. This causes no issues with the way pomf currently has the watchdog set up to pick maps, and is better for testing.