Changes cult sacrifice objectives to not say MODE
Currently, cult sacrifice objective goes solely off of
`sacrifice_target.assigned_role` which, if special roles like ERT members
are in play, will make the objective read "Sacrifice Big McLargehuge, the
MODE."
This changes it to use the target's special_role (which seems to be
designed specifically for this) by ripping off someone else's work in
code\datums\mind.dm
I'm at a loss regarding how to properly test this, so if anyone can assist
there, that'd be super. I can't seem to select the game mode datum using
SDQL and it doesn't appear in the cultists' minds.