Hyperzine, Nuka Cola & Synaptizine now act as drugs for Diona These are highly addictive and hallucinogenic drugs for diona - think sensory overload mixed with skooma. They will increase, very minorly but enough to be annoying: - Hyperzine/Nuka Cola now only brings you, as a diona, to the normal running speed of a human - not faster than one. (If you want to go faster, drink hyperzine AND nuka cola.) - druggy variable (I believe this is the colourful lights of space drugs) - increases brain damage (very slowly) - adds hallucinations - makes dizzy - chance of making confused (ie moving in wrong direction) - chance of vomiting - could twitch/moan/drool/dribble
These are all capped to reasonable amounts - so this can be treated by charcoal, a bit of antitoxin, time, and alkysine by medbay if they OD.

The intention is to not take away the ability for people to be fast diona, but considering that's one of their racial cons - if you want to be a fast diona, you have to take other, somewhat annoying but still absolutely playable cons in return.

Note to self: Hyperzine doesn't seem to be increasing jitteriness anymore - not a bug with this PR though, could reproduce on live.