Fixes #11180 (#12877)
Fixes #11180
Warning: I have tested this, and I do know this works, but I don't know if this has unintentional side effects. It !!shouldn't!! but that doesn't mean it doesn't. I think Clusterfack would be the one to ask as to why this uses this system

The code is five years old, I git blamed. - I don't understand why it's here, because any time where a mob's loc will be inside of something, the code above works fine?

My suspicion is it's to do with mechs? But it really shouldn't be.