QoL for Snowmap: Wood walls constructed faster, but are far easier to break (#13153)
One of the bonuses of snowmap is the greater ability of people to build little huts in the snow. I'd like to encourage that behaviour; so I have halved the construction time of wooden girders. To balance this, you can break through wooden walls with an axe with two swings - one to break the wall, the other to break the girder