Incredibly huge plasma nerf
comments out section that causes jumpsuits, gloves, and shoes to become plasma-fied. Ive seen it proposed multiple times but ever actually acted on. The reasoning is that plasma already such an easy win for anyone using it that having it toxify clothing was irrelevant since if it is going to kill someone it will burn someone to death. This was also silly because naked people could come out unscathed meaning that wearing these objects decreased your chances of survival.