Sweet Machine Buff
Buffs the Sweet Machine so that it dispenses a bag of sweets instead of
just a single sweet.

The same holds for emagged sweet machines, you now get a bag filled with
ten sweets (either normal or strange ones).

Expanded the list of possible chemicals for strange sweets to include a
couple more interesting effects;.