Should be it!
There's an issue with them not reporting a second, non specialist damage
type, but that's because it won't call the procs to heal specific damage
types and thus, play the lines, it the patient already has tricord in
them, which is uses for every injury type except the specialist one. Not
ideal, but would be very complex to fix. Interestingly, does make the
Mysterious Medbot work exactly as intended, since it has a different
chem for every ailment.