Talking medbots (#14775)
* Added the sounds!

Added the new folder, and the relevant sound files that we're going to
be using.

* A start

Some voice lines swapped out, but not the big ones.

* Mostly there!

Just need to add a world-time based timer to the speech section to stop
it repeating itself, since /vg/ bots do a double injection if there's a
damage type outside their specialist one, and we're good to go!

* Should be it!

There's an issue with them not reporting a second, non specialist damage
type, but that's because it won't call the procs to heal specific damage
types and thus, play the lines, it the patient already has tricord in
them, which is uses for every injury type except the specialist one. Not
ideal, but would be very complex to fix. Interestingly, does make the
Mysterious Medbot work exactly as intended, since it has a different
chem for every ailment.

* Added a medbot SFX channel, and assigned their dialogue to it.

What the title said. Seems to cause them to skip messages rather than
overlay them if two trigger at once, but that's much better than a
spamming anyway, so I'll take it.