Cult Sacrifice & Raise "resurrection charges"
Specifically:
Sacrificing someone now adds a resurrection charge.
Attempting to convert someone that is cultbanned strips and deletes them, adds a resurrection charge and tells you a runeword.
Using the Raise rune will attempt to use a corpse or a resurrection charge.
Using the Raise rune without a corpse on it will tell you the amount of resurrection charges available.
Also added a message that a conversion isn't working when someone isn't knocked down and fixed some gender issues.