Adds "bridge officer" alt title job to IAA
Adds "Bridge Officer" to IAA's alt titles. Bridge Officer is a junior command position below the Heads that exists to coordinate efforts from the bridge using consoles and command headset. It is a simple job that functions differently from IAA or Lawyer but fits within the administrative branch of the station.