Implements cantpass_flags
I spent 2 hours trying everything I could think of in regards to movement code.

This is what the BYOND documentation says: 
>2.newloc.Enter(src) is called for any turfs or areas that may be entered for the first time, or the container if moving into an obj or mob. neighbor.Cross(src) is called for any movable atoms that may be in collision with this object if the move fully succeeds. If any of these return 0 (failure), then a slide can be cut short but a jump will fail completely.

>3.If any obstacles were encountered via Enter() or Cross() failing, then src.Bump(obstacle) will be called for each of them
And so, we simply can't modify a chair's to_bump because it happens after the Cross() fails, and there is no reciprocal proc when an atom is being Cross()ed to have the moving atom check for shit.
Necessarily you gotta do this with the stationary item, and in our code the only things that do that are density and passflags. I had to add cantpass_flags, unless I misunderstand something here.