Update, Security.dm, Detective's Closet
changes the code for the item list that will spawn in the detective's closet. The changes will have the closest include a camera bug box, and the handtv. These items were mapped into Bagel station so to avoid duplication they would have to be mapped out if this were to be merged.