NanoUI object focus improvement.
Previously if you clicked on an object and you already had the NanoUI open, two things could happen:
* The UI only updates
* The UI reopens with a white BYOND flash.

The latter did always move it to the foreground, but the flash and UI moving was pretty ugly.
The former only updates the UI, so if you had multiple stacked windows open it gets kinda annoying.

With this change, all UIs always do the former, and the UI for an object always moves to the front when you interact with said object.