Removes unnecessary delays on handled attackby.
You know how when wrenching pipes, if you misclick and hit one you already placed, you can't place pipes for a moment?

Yeah some genius decided that if an attackby is handled it MUST generate a cooldown. Every object that DIDN'T have a cooldown on attackby was literally misusing attackby.

This dumb cooldown is finally removed. There's still cooldowns on attacking and stuff as far as I can tell.

If anything becomes exploitable and spam clickable we can fix that later.