Ghetto Surgery no longer bashes people
And a bunch of other things sorted out. Like being unable to work on people in suits, unless people complain about it to be able to work throug suits, and being able to work on concious people using pain meds.

Also autopsy was moved to be a surgery so any valid surfaces that surgery considers will work, albiet with failure rates.

Additionally, you need to be on the harm intent to actually stab people in the eyes.