Online edits 3 END
That should be about it.
CAN_DO_SURGERY_ON_DISARM_GRAB_INTENT changes if surgery can happen even if on GRAB or DISARM and also changes if forks or screwdrivers should stab people in the eyes. Generally if it's switched on, you have to stab someone in the eyes on the 'HURT' intent, otherwise any intent other than 'HELP' will just stab people in the eyes.