Snow-map Salvage Project PR#4 - clothing tweaks edition! (#17486)
Adds the security ushanka - it goes extremely well with the warden winter coat. (should have been in the winterjacket PR)
Rejiggles a couple of heat conductivities with neorussian outfits; fixes some wintercoats not properly displaying stuff under it,
adds ski mask (literally the same as a balaclava but a different name)
changes balaclava to have the correct flags for covering the full head except eyes.