Adds craftable lids for beakers, bottles and beakers.
As it says on the tin, I've always wondered where spessmen found lids for their beakers, bottles and buckets. This PR will remove magical lids that appear out of thin air to cover the beaker. If you wish to put a lid on your reagent container of choice, you're going to have to craft a lid.
Lids are tiny items that can be crafted with glass.

:cl:
	*rsadd: Adds craftable lids for your beakers. They can be crafted using glass.