Remove flashbangs from the game.
Flashbangs are by far the worst and most overpowered object of the game,
offering a 10*10 stun of over 15 seconds to most foes, while being
abundant and virtually inoffensive against the security officers who use
it.

Most importantly, they stun cyborgs, who are by far in need of buffs in
the current ss13 meta.

This change aims to restore an attempt at balance in the ss13 ecosystem
by simply removing flashbangs from the game.

:cl:

- tweak: Flashbangs have been removed from the game.