The last word in infiltration technology
"Doesn't seem much use to me, but Creed was adamant that you'd want it on the list. 'The last word in infiltration technology'...or so he said."

You can now craft cardboard boxes out of 5 cardboard "sheets". When you use the box, your sprite will disappear and will instead be replaced by the box, much like a chameleon projector.