O2 First-Aid Kits now contain 2 Inaprovaline pills instead of 1 Inaprovaline syringe
After a bit of discussion in the Inaprovaline metabolism PR (#18016), I've decided to swap around the contents of the O2 First-Aid kit in order to encourage the use of Inaprovaline, along with standardising the usage of pills over syringes. Each Inaprovaline pill contains 30 units of Inaprovaline and should benefit people that didn't like having to triple-inject someone to give them 15 units of Inaprovaline.

:cl:
	*tweak: O2 First-Aid kits now contain 2 Inaprovaline pills and 3 Dexalin pills, replacing the old inaprovaline syringe and 4 Dexalin pills.