Adds "Brotherhood of the Yellow Sign" Religion (#18159)
* Adds "Brotherhood of the Yellow Sign" Religion

I know there's a religion for lovecraft shit already, but the religion name (which it shows to converts) is "Esoteric order of Dagon". People see this and assume I worship Dagon, which makes sense, because Dagon (a specific entity) appears to be the object of worship here. This, however, is not supposed to be an unatomic commit, so I'm fixing the problem by adding another religion. I play chaplain often, so it would definitely see use. Any criticism and further changes are welcome, and I'll probably open an issue on the whole "all lovecraft things are dagon" bit.

* dude capitalization lmao

apparently you capitalize the "the" in "The Yellow Sign"