Staff of Swapping
Adds the Staff of Swapping. Fires a projectile that, when it collides with a mob, swaps the position of the firer and the mob.

Moves the staff of change to its own subtype so all staffs don't keep inheriting the staff of change's attack_hand. Try attack_handing a staff of animation, I dare you.