Turning into a ghoul now sets radiation, rad_tick & cloneloss to 0. Slimepeople get cloneloss set to 0.
You know, for species that are supposed to be immune to the effects of what they're supposed to be immune to, they still get fucked because of what they had before switching.

Ghouls:
Managed to become a ghoul? You're a lucky guy, now enjoy cloneloss caused from...radiation (which you're immune to) to destroy you. Or your pre-existing rad_tick and radiations.

Slimepeople:
Even if you're immune to clone damage, if you had clone damage before turning into the master race, you'll still somehow keep the clone damage even if you're immune to it.