Removes the ablative vests from the Experimental Energy Gear crate (#18781)
* Removes the ablative vests from the Experimental Energy Gear crate

For these reasons:
A) Screw having to sort out the ablative vests from a crate that's used to get guns. The current blob "meta" on the crew's side is to amass eguns as fast as possible, not sift through vests.
B) There's already another crate that can get you Ablative Vests, so if a cargonian really wants to rebel, then he's going to pay for it instead of getting it for "free".
C) The primary members of the Cargo metaclub have allowed this.