Makes space in the HoS locker for everything to fit. (#19453)
* Makes room for a debit card in the HoS locker.
The Discord pointed out a lone flashbang was in the locker despite a box of them already in it. The lone flashbang was removed.
Also reserved an object when a mapper decides to add another locker to HoS dedicated to attire.

* Just makes room in the HoS locker.
Flashbang still removed, but moves the lawgiver magazine into the lockbox instead.