Role datums objective fixes
Fixes objectives not functioning properly. This was due to them looking for a mind that had not been assigned at that point.

Also fixes steal objectives always showing as successful by removing a just be yourself joke that was being inherited.

Fixes the vampire objective whitespace, switching it from spaces to tabs.