Ghetto spacesuit helmets now have flashbang protection (again?)
According to pic related and the word of older players, ghetto spacesuit helmets once protected you from bangers, but sanity fucked it all up and they got dropped to 0 eye_protection, which is imperative in order to not get megafucked by a flashbang (even if you have ear protection). While flashbangs are the main problem (Huge-ass range that IGNORES WALLS & 1 second fuse time), the current statu quo is that unless you've got a hardsuit (and softsuit?) helmet or a security helmet, get fucked because you're going to eat the (nigh) full force of the banger. Thus, I propose this change so basically anyone that doesn't get a free gibs as part of their job can have an amount of protection against bangerbois (Read: greyshits, cultists, revs, etc).
Also Shifty said something about maybe looking into making it so flashbangs can't work through walls, so that's a thought for the future.