Adds potion bags for wizards in magi-vends.
This is a blatant buff to potions and I don't really expect it to get used by anyone else than me if I ever rolled wizard, but hey. This brown paper bag can be attached to your belt for quick access or strapped to your head for tactical /no-robes/ action and halloween (hobo) costume parties. It can hold 50 potions through means unknown to mere mortals, but NT experts suggest that it is based on the principle of "It's magic, we don't have to explain shit".

Sprites courtesy of RealestEstate.