Runtime hunting: Assert sound loc
Because of things that delete themselves on damage/death, such as spiderlings, the location of the attack playsound would be null by the time it would reach on_attack, as they would have deleted themselves and be nullspaced.

Now, on_attack is called before the damage is dealt rather than after.