Z levels only process if there's a player on that zlevel
Just a silly idea I had, when I heard that TG or somewhere had processing only happen when a player was on it.

Needs to be enabled in config, and when it is, creates a list in the map subsystem that holds however many zlevels there are amount of lists.

Each list holds a reference to a mind datum, which is moved around during z level transition.

Currently object, machinery, and mob processing require there to be a mob on the zlevel, should the list of lists be initialized.

Also, moves the on_z_transition event to on_z_transition proc (atom level), which is called by a players move and forceMove, rather than everywhere.