Essence printer
Adds the Essence Printer, sprite and idea courtesy of Megautham on discord

You click it, it binds your essence to it, not unlike a phylactery.

On death, it creates a duplicate of your body over 30 seconds to a minute. Problem is, ANY ghost can take it over, and it halves the health of the clone.