Weeb antag is now more likely to roll.
Because the old values made it require as much threat/population range as Nuke Ops. Weebs are still as unlikely to roll on lowpop, but they're now as likely to roll as a Traitor in the 15-19 population range and higher. In addition, they weigh a tiny bit more than wizards, vampires and grinches.