Adds the Potion Launcher
Long overdue (I said I'd do this about 2 months ago).

Coming at relatively cheap cost (5 spell points), the potion launcher is meant to be used for rapid firing of potions. It can hold up to 5 potions in its loader and also comes with a silly new sound effect when it fires. It also can be attached to the suit slot of wizard robes and gem hardsuits.