Makes vampire less likely to fire at roundstart (especially on lower populations).
-Weight changed from 3 to 2.
-Cost changed from 18 to 25
-Theat cost changed from (80,60,40,20,20,10,10,10,10,10) to (80,60,50,30,20,10,10,10,10,10)

I understand that Vampire more or less rolls often because apart from Traitor, it's just about the only other antag that has a "low" cost to fire, so its current weight doesn't matter as much in the grand scheme of things. That said, I'm not a big fan of every round or two in a row containing 1+ vampire instead of something else.