Mug overlays (#21706)
* mug overlays

Adds reagent overlays to mugs. Also changes said mugs name in a fashion similar to drinking glasses.

* CHOCCY MILK

Adds a special entry for hot coco, because it's kind of dumb that it didn't have one in the first place.

* blood desc

Added a description to the blood reagent, since it was lacking one. Stuck with the "it's totally a tomato" scheme that the rest of the descriptions has.

* blood mugs, remove overlay when empty